local-conf.in: remove image-mklibs from USER_CLASSES (build 352 part 1/3)
:Release Notes:
It was never really enabled, because MKLIBS_OPTIMIZED_IMAGES
wasn't set in our builds.

:Detailed Notes:
It's broken for a while and now finally removed from oe-core/master
as well in:
https://git.openembedded.org/openembedded-core/commit/?id=908df863b419d1cad7317153101fc827e7e3a354
which breaks the builds which still inherit it with:
ERROR: ParseError in configuration INHERITs: Could not inherit file classes/image-mklibs.bbclass

This whole file should probably be refreshed to match the current samples in:
https://git.openembedded.org/openembedded-core/tree/meta/conf/local.conf.sample
+
https://git.openembedded.org/openembedded-core/tree/meta/conf/local.conf.sample.extended
as it was created from them, but in 2012 based on Yocto 1.3 Danny
and wasn't updated much in build-* repos since then.

:Issues Addressed:
[PLAT-142565] CCC: Upgrade to latest revisions from Yocto 3.1 Dunfell

Change-Id: I36310d70a7719f0ba1c4eb464e309a14bab27ab3